requirements.txt
Flask
pandas
numpy
scikit-learn
pymysql
python-dotenv
